# api/openapi.yaml
openapi: 3.0.3
info:
  title: Attachra REST API
  version: "1.0.0"
  description: |
    Attachra's admin/automation REST API (US-8.1, epic E8). This document is
    the **source of truth** for the API contract (ADR / API-first principle):
    the HTTP handlers implementing each operation below (T-8.1.2..T-8.1.7)
    must match it, not the other way around.

    ## Scope

    This spec covers only the versioned, token-authenticated admin API under
    `/api/v1`. Three other HTTP surfaces already implemented by
    `internal/adapters/http` are intentionally **out of scope** here, and are
    listed for completeness rather than documented as paths below:

    - `GET /p/{token}` and `POST /p/{token}/d/{ref}` — the anonymous
      recipient-facing package/download pages (US-6.2). These are not JSON
      resources, carry no Bearer auth, and are documented by their own
      package (`internal/adapters/http/handler.go`,
      `docs/architecture/package-page-decision.md`).
    - `GET /metrics` — Prometheus text exposition format, not JSON; described
      by the metrics library itself, not by an OpenAPI document.
    - `GET /healthz` and `GET /readyz` — liveness/readiness probes
      (`internal/adapters/http/health.go`). Neither is mounted under
      `/api/v1` (`internal/adapters/http/server.go`), so including them as
      `paths` here — which inherit this document's `servers: [{url:
      /api/v1}]` — would misrepresent their real mount point as a contract
...

    ## Versioning

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /messages/{messageId}:
    get:
      operationId: getMessage
      summary: Get a single message by ID
      tags: [Messages]
      x-required-role: [admin, viewer]
      parameters:
        - $ref: '#/components/parameters/MessageIdPath'
      responses:
        '200':
          description: The message.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /links:
    get:
      operationId: listLinks
      summary: List personal download links
      tags: [Links]
      x-required-role: [admin, viewer]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: message_id
          in: query
          schema: { type: string }
        - name: recipient
          in: query
          description: 'Exact recipient match, case-insensitive with SMTP angle brackets ignored (ATR-293).'
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/LinkStatus' }
        - name: from
          in: query
          description: Inclusive lower bound on `created_at`.
          schema: { type: string, format: date-time }
        - name: to
          in: query
          description: Exclusive upper bound on `created_at`.
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: A page of links.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LinkList' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /links/{linkId}:
    get:
      operationId: getLink
      summary: Get a single link by ID
      tags: [Links]
      x-required-role: [admin, viewer]
      parameters:
        - $ref: '#/components/parameters/LinkIdPath'
      responses:
        '200':
          description: The link.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Link' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /links/{linkId}/revoke:
    post:
      operationId: revokeLink
      summary: Revoke a single link
      description: >-
        Mirrors `link.Engine.Revoke`. Refuses with `409 held` if the link is
        currently under legal hold (ATR-233/ATR-257) — the hold must be
        explicitly lifted via `POST /links/{linkId}/unhold` first.
      tags: [Links]
      x-required-role: [admin]
      parameters:
        - $ref: '#/components/parameters/LinkIdPath'
      responses:
        '200':
          description: The link, now revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Link' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: The link is under legal hold and cannot be revoked yet.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: { code: held, message: "link is under legal hold" } }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /links/revoke-by-message:
    post:
      operationId: revokeLinksByMessage
      summary: Revoke every link belonging to one message (cascade)
      description: >-
        Mirrors `link.Engine.RevokeMessage`: revokes every link of the
        message except any currently under legal hold, which are skipped
        and reported back via `held`, not treated as a hard failure — a
        partial revoke is still reported as `200`, with `held > 0`
        signalling that at least one link needs an explicit unhold first.
      tags: [Links]
      x-required-role: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RevokeByMessageRequest' }
      responses:
        '200':
          description: Revoke outcome (possibly partial, see `held`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RevokeByMessageResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /links/revoke-by-sender:
    post:
      operationId: revokeLinksBySender
      summary: Revoke every link belonging to every message from one sender (cascade)
      description: >-
        Mirrors `link.Engine.RevokeSender`, after resolving the sender's
        message IDs via `store.MetadataStore.ListMessagesBySender`. Like
        the by-message variant, held links are skipped and reported, not
        treated as a hard failure.
      tags: [Links]
      x-required-role: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RevokeBySenderRequest' }
      responses:
        '200':
          description: Revoke outcome (possibly partial, see `held_messages`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RevokeBySenderResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api-tokens:
    get:
      operationId: listApiTokens
      summary: List API tokens
      description: Never returns a secret or its hash — only issuance metadata.
      tags: [API Tokens]
      x-required-role: [admin]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: A page of API tokens.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiTokenList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createApiToken
      summary: Create a new API token
      description: >-
        The response's `secret` field is the only time the raw bearer
        token is ever shown; only its hash is persisted thereafter
        (SR-130-2). There is no "show secret
        again" endpoint — losing it means issuing a new token.
      tags: [API Tokens]
      x-required-role: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApiTokenCreateRequest' }
      responses:
        '201':
          description: The token was created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiTokenCreateResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >-
        An Attachra API token secret, shown once at creation time
        (`POST /api-tokens`). Never accepted as a query parameter
        (SR-130-2); compared server-side in constant time against the
        stored SHA-256 hash.
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      description: Opaque pagination cursor from a previous page's `next_cursor`. Omit for the first page.
      schema: { type: string }
    MessageIdPath:
      name: messageId
      in: path
      required: true
      schema: { type: string }
...
    LinkIdPath:
      name: linkId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Malformed request (bad query parameter, invalid JSON body, etc).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: bad_request, message: "invalid value for query parameter \"limit\"" } }
    Unauthorized:
      description: Missing, malformed or invalid Bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: unauthorized, message: "missing or invalid bearer token" } }
    Forbidden:
      description: The token's role is not permitted to call this operation (see `x-required-role`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: forbidden, message: "this token's role is not permitted to perform this action" } }
    NotFound:
      description: No resource exists with the given identifier.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: not_found, message: "no resource found for the given identifier" } }
...
    TooManyRequests:
      description: Rate limit exceeded (SR-130-5).
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: rate_limited, message: "too many requests, slow down" } }
    InternalError:
      description: An unexpected server-side failure. The response never includes internal detail (no stack trace, driver error text, or file path — SR-130-1).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: internal, message: "an internal error occurred" } }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable error identifier.
              enum:
                - bad_request
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - held
                - invalid_policy
                - validation_failed
                - payload_too_large
                - rate_limited
                - gone
                - internal
            message:
              type: string
              description: Safe, human-readable message. Never includes internal detail (SR-130-1).
            details:
              type: array
              description: Structured, per-field issues. Populated only by policy validate/reload failures.
              items: { $ref: '#/components/schemas/ValidationIssue' }
    Action:
      type: string
      description: One of the three policy actions (`policy.Action`).
      enum: [pass, replace, block]
    LinkStatus:
      type: string
      description: Link lifecycle state (`store.LinkStatus`).
      enum: [active, revoked, expired]
    Role:
      type: string
      description: >-
        One of Community's three fixed roles (ADR-015). The model is
        intentionally left extensible (Enterprise Identity Pack may add
        custom roles) but Community itself implements exactly these three.
      enum: [admin, viewer, auditor]
...
    Message:
      type: object
      required: [id, queue_id, sender, created_at]
      properties:
        id: { type: string, description: "store.Message.ID (milter queue ID-derived message identifier)." }
        queue_id: { type: string, description: "store.Message.QueueID — the milter queue ID." }
        sender: { type: string, description: "store.Message.Sender — the envelope-from address." }
        created_at: { type: string, format: date-time }
        recipients:
          type: array
          items: { type: string }
          description: Distinct recipients derived from this message's links (not a stored column on `store.Message` itself).
        attachment_count:
          type: integer
          description: Number of `store.Attachment` rows belonging to this message.
        status:
          allOf: [{ $ref: '#/components/schemas/Action' }]
          type: string
          nullable: true
          description: >-
            Aggregated message-level decision. See the `status` filter's
            implementation note on `GET /messages` — not backed by a
            persisted column today.
...
    Link:
      type: object
      required: [id, message_id, attachment_id, recipient, expires_at, max_downloads, downloads, status, hold, created_at]
      properties:
        id: { type: string }
        message_id: { type: string }
        attachment_id: { type: string }
        recipient: { type: string }
        expires_at: { type: string, format: date-time }
        max_downloads: { type: integer, minimum: 0, description: "0 means unlimited (store.Link.MaxDownloads)." }
        downloads: { type: integer, minimum: 0 }
        status: { $ref: '#/components/schemas/LinkStatus' }
        hold: { type: boolean, description: "Legal hold flag (ATR-233/ATR-257). While true, revoke and retention deletion are refused." }
        hold_set_by: { type: string, nullable: true }
        hold_set_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
      description: Deliberately excludes `token_hash` — never exposed via the API.
    LinkList:
      allOf:
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/Link' }
        - $ref: '#/components/schemas/PageMeta'
    RevokeByMessageRequest:
      type: object
      required: [message_id]
      properties:
        message_id: { type: string }
    RevokeByMessageResult:
      type: object
      required: [revoked, held]
      properties:
        revoked: { type: integer, description: Number of links actually revoked. }
        held: { type: integer, description: Number of links skipped because they are under legal hold. }
    RevokeBySenderRequest:
      type: object
      required: [sender]
      properties:
        sender: { type: string, description: 'Matched case-insensitively, with SMTP angle brackets ignored (ATR-293).' }
    RevokeBySenderResult:
      type: object
      required: [revoked, held_messages]
      properties:
        revoked: { type: integer, description: Total links revoked across every message from this sender. }
        held_messages: { type: integer, description: Number of messages that had at least one link skipped due to legal hold. }
...
    ValidationIssue:
      type: object
      required: [path, message]
      properties:
        path: { type: string, description: "JSON-pointer-like location, e.g. \"rules[2].then\" or \"default\" (policy.ValidationError.Path)." }
        rule_name: { type: string, description: "Name of the offending rule, when known." }
        message: { type: string }
...
    ApiToken:
      type: object
      required: [id, name, role, created_at]
      properties:
        id: { type: string }
        name: { type: string, description: Operator-chosen label for this token. }
        role: { $ref: '#/components/schemas/Role' }
        created_at: { type: string, format: date-time }
        last_used_at: { type: string, format: date-time, nullable: true }
      description: Never includes the token secret or its hash.
    ApiTokenList:
      allOf:
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/ApiToken' }
        - $ref: '#/components/schemas/PageMeta'
    ApiTokenCreateRequest:
      type: object
      required: [name, role]
      properties:
        name: { type: string }
        role: { $ref: '#/components/schemas/Role' }
    ApiTokenCreateResponse:
      allOf:
        - $ref: '#/components/schemas/ApiToken'
        - type: object
          required: [secret]
          properties:
            secret:
              type: string
              description: >-
                The raw bearer token. Shown exactly once, at creation
                time; only its hash is persisted (SR-130-2). There is
                no way to retrieve it again.
    PageMeta:
      type: object
      required: [next_cursor]
      properties:
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page, or null if this is the last page.